requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
yfinance>=0.2.36
matplotlib>=3.8.0
python-dotenv>=1.0.1
google-generativeai>=0.4.0
plotly>=5.20.0
requests>=2.31.0
protobuf~=5.29.5
dotenv~=0.9.9